raft: short-circuit debug logging when not enabled
This commit creates a new DebugLogger interface which Logger
implementations can choose to implement. The new interface
exposes a method for the `raft` package to check whether debug
logging is enabled or whether a debug log call would be ignored.
Callers of Logger.Debug and Logger.Debugf now check whether debug
logging is enabled before calling the debug methods. This allows
them to avoid memory allocations when passing arguments through
the debug calls if the calls will be no-ops.